add dom manipulation to jest and create mocks and test file to separates test and mocks from current project
