+++ __mocks__/ContainerMock.js
@@ -1,0 +1,5 @@
+const ContainerMock = document.createElement('div');
+ContainerMock.id = 'container';
+document.body.appendChild(ContainerMock);
+
+export default ContainerMock;
